Fixed clang warnings

+++ trunk/source/newtonian/test_2d/sb_edge_repellant.cpp
@@ -43,10 +43,10 @@ vector<Vector2D> SBEdgeRepellant::calcAllVelocities(Tessellation const& tess,
 {
   vector<Vector2D> result = naive_.calcAllVelocities
     (tess,cells,time,cevolve);
-  for(int i=0;i<total_specials_;++i)
+  for(size_t i=0;i<static_cast<size_t>(total_specials_);++i)
     result[i] = Vector2D(0,0);
   for(size_t i=static_cast<size_t>(total_specials_);i<result.size();++i){
-    const Vector2D mp = tess.GetMeshPoint(i);
+    const Vector2D mp = tess.GetMeshPoint(static_cast<int>(i));
     const Vector2D vm = cells[i].Velocity;
     if(should_be_decelerated(mp,vm)){
       const Vector2D rhat = mp/abs(mp);
